CORE-2083 Refactor request/response types into `@/app/api/types`
